inventario_host/requirements.txt
diff-match-patch==20121119
Django==1.11.7
django-admin-list-filter-dropdown==1.0.1
django-filter==1.1.0
django-import-export==0.5.1
djangorestframework==3.7.3
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
mysqlclient==1.3.12
odfpy==1.3.5
openpyxl==2.4.9
psycopg2==2.7.3.2
pytz==2017.3
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
unicorn==1.0.1
xlrd==1.1.0
xlwt==1.3.0
